requirements.txt
asgiref==3.3.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
coverage==5.3
cryptography==3.2.1
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.1
idna==2.10
Pillow==8.0.1
pycparser==2.20
Pygments==2.7.2
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
stripe==2.55.0
urllib3==1.26.2
gunicorn==20.0.4
whitenoise==5.2.0
